Move worker related files from data to lib folder, so that we do not use require('self') anymore.
